server/requirements.txt
Django==4.2
psycopg2-binary>=2.8
pytest==8.2.0
pytest-cov==5.0.0
pytest-django==4.8.0
django-extensions==3.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt
django-extensions
django-cors-headers==4.3.1
django-admin-sortable2==2.0.3
django-filter==23.5
django-jazzmin
pre-commit
shortuuid==1.0.13
